requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
click==8.3.2
    # via uvicorn
fakeredis==2.35.1
    # via -r requirements.in
fastapi==0.135.3
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
iniconfig==2.3.0
    # via pytest
packaging==26.1
    # via pytest
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.20.0
    # via pytest
pytest==9.0.3
    # via -r requirements.in
python-dotenv==1.2.2
    # via -r requirements.in
redis==7.4.0
    # via
    #   -r requirements.in
    #   fakeredis
sortedcontainers==2.4.0
    # via fakeredis
sqlalchemy==2.0.49
    # via -r requirements.in
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.44.0
    # via -r requirements.in
